privacy multi-tenancy controller
